fix: dove turns up in activeTab with valid URL
The dove now turns up without doing a no show by using chrome.alarms, which sometimes don't get cancelled when the background.js gets terminated. However, this is very flaky and I've spent way too long on this, so if there's an error that the user can fix, I've made it look nice with a notification for them.